first commit: project files, initial `@supply_args` implementation
Here I'm adding the initial implementation of the `@supply_args`
decorator.

Previously it was a part of the `contextvars-extras` repo:
https://github.com/vdmit11/contextvars-extras

...but I decided to extract it to a separate Python package,
and a separate git repo.

So its not an empty commit. There is a lot of code already written.

The project files and scripts are also copied from the
`contextvars-extras` package. I tried to clean up all references
and re-name everything to `supply-args`.

So, this commit contains a self-sufficient Python project,
with project files, code, tests, and even some documentation.