drop support for `@suply_args(dict)` form of the decorator
Problem: when you pass `dict` to `@supply_args`, like this:

    @supply_args(
        {'source': config, 'names': ['locale', 'timezone']}
    )

it converts it implicitly to `SupplySpec()` object.

That works, but it completely prevents from using `dict`
as a source of values, which is a feature I would like to have.

So, I decided to drop this way of configuring the decorator,
and replace it with explicit `SupplySpec` instances.
That is, instead of this:

    @supply_args(
      {'source': config, 'names': ['locale', 'timezone']}
    )

you now have to do this:

    @supply_args(
       SupplySpec({'source': config, 'names': ['locale', 'timezone']})
    )

This is slightly longer, but in exchange, I can use `dict` as source,
like this:

    config = {}

    @supply_args(config)

(this is not yet implemented, will be added in the next commit)